+++ sydtest-persistent-sqlite/src/Test/Syd/Persistent/Sqlite.hs
@@ -70,5 +70,7 @@ runSqliteTest = runPersistentTest
 -- | Test that the given migration succeeds, when applied to the current database.
 --
 -- See 'Test.Syd.Persistent.migrationsSucceedsSpec" for details.
-sqliteMigrationSucceedsSpec :: FilePath -> Migration -> Spec
-sqliteMigrationSucceedsSpec = migrationsSucceedsSpecHelper connectionPoolSetupFunc
+sqliteMigrationSucceedsSpec :: FilePath -> Migration -> TestDef outers void
+sqliteMigrationSucceedsSpec fp migration =
+  setupAround (connectionPoolSetupFunc (pure ())) $
+    migrationsSucceedsSpecHelper fp migration